Reset path

+++ app/utils/index.js
@@ -4,11 +4,11 @@ const path = require('path');
 const isDev = require('electron-is-dev');
 const copyFile = require('fs-copy-file');
 const mkdirp = require('mkdirp');
+const fs = require('fs');
 
-export const userDataPath = process.resourcesPath;
-// (electron.app || electron.remote.app).getPath(
-//   'userData'
-// );
+export const userDataPath = (electron.app || electron.remote.app).getPath(
+  'userData'
+);
 export const desktopPath = (electron.app || electron.remote.app).getPath(
   'desktop'
 );
@@ -17,29 +17,34 @@ export const defaultHosts = path.join(userDataPath, '/default.hosts');
 export const usersHosts = path.join(userDataPath, '/user.remote.hosts');
 
 export const initUntrack = async () => {
-  mkdirp(`${desktopPath}/untrack`, err => {
-    if (err) console.error(err);
-    else {
-      copyFile(`/etc/hosts`, `${desktopPath}/untrack/hosts`, error => {
-        if (error) throw error;
-        copyFile(
-          `${getScriptsPath}/restore.sh`,
-          `${desktopPath}/untrack/restore.sh`,
-          errors => {
-            if (errors) throw errors;
-          }
-        );
-      });
-    }
-  });
+  if (fs.existsSync(`${desktopPath}/untrack/hosts`)) {
+    console.log('already exist');
+  } else {
+    mkdirp(`${desktopPath}/untrack`, err => {
+      if (err) console.error(err);
+      else {
+        copyFile(`/etc/hosts`, `${desktopPath}/untrack/hosts`, error => {
+          if (error) throw error;
+          copyFile(
+            `${getScriptsPath}/restore.sh`,
+            `${desktopPath}/untrack/restore.sh`,
+            errors => {
+              if (errors) throw errors;
+            }
+          );
+        });
+      }
+    });
+  }
 
-  copyFile(`/etc/hosts`, `${defaultHosts}`, err => {
+  copyFile(`${desktopPath}/untrack/hosts`, `${defaultHosts}`, err => {
     console.log('copied', defaultHosts);
     if (err) throw err;
   });
 };
 
 export const copyScripts = async () => {
+  console.log(`${getScriptsPath} ${userDataPath}`);
   copyFile(
     `${getScriptsPath}/functions.sh`,
     `${userDataPath}/functions.sh`,
